refactor(auth-service): rename files
