+++ apps/auth-service/src/interfaces/user.interface.ts
