+++ apps/auth-service/src/models/User.model.ts
