+++ apps/auth-service/src/routes/health.route.ts
