+++ apps/auth-service/src/routes/login.route.ts
@@ -1,8 +1,8 @@
 import passport from '../middleware/passport.init';
 import PromiseRouter from 'express-promise-router';
-import { loginController } from '../controllers/loginController';
+import { loginController } from '../controllers/login.controller';
 import { validate } from '@acua/microservice-shared';
-import { loginSchema } from '../validators/login';
+import { loginSchema } from '../validators/login.validator';
 
 const router = PromiseRouter();
